requirements.txt
instagrapi==1.17.0
openai==0.27.0
Pillow==9.4.0
stability_sdk==0.3.2
